% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exa.R
\name{exa_jdbc}
\alias{exa_jdbc}
\title{Exasol JDBC Connection}
\usage{
exa_jdbc(file = NULL, ...)
}
\arguments{
\item{file}{connection file (defaults to '~/.exa' if no other arguments are provided)}

\item{host}{host, e.g. 'localhost:8563'}

\item{schema}{schema (default: 'public')}

\item{user}{user (default: 'sys')}

\item{password}{password}
}
\description{
Set connection parameters in `file` or pass them as arguments to this function.
}

