% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exa.R
\name{exa_query}
\alias{exa_query}
\title{Query Exasol database}
\usage{
exa_query(.con, .sql, ...)
}
\arguments{
\item{.con}{connection}

\item{.sql}{sql query}

\item{...}{use infuser to provide arguments to the sql query}
}
\description{
Query Exasol database
}

